Create TATypes.hpp
This file has list() fully functional and so is evaluate() fully functional. However, a modification was implemented which is seperating each of TAPlus, TAMinus, TAMultiply, TADivide into two classes each concerned with either TAInt or TADouble.

Other than that, for problem 3, a super class was implemented to have the TATypes as children so that TAPair would be functional. Now what's left are some tweaks to satisfy the requirements of the assignment.

N.B.: The late submission of this assignment was excused for evaluate() to be fully functional. This monday we had Dr. Fadi explaining virtual functions, which helped in successfully implementing evaluate's recursion.